problem not fetching solved
