+++ src/app/App.js
@@ -5,7 +5,7 @@ import 'materialize-css/dist/css/materialize.css'
 import M from 'materialize-css'
 import './App.css';
 
-import { Header } from './partials/Header';
+import Header from './partials/Header';
 import { Footer } from './partials/Footer';
 import { Feed } from './partials/Feed';
 import { PostContent } from './partials/PostContent';
@@ -13,7 +13,7 @@ import { Profile } from './pages/userProfile/Profile';
 import { PeoplePage } from './pages/People/PeoplePage';
 import { UserPage } from './pages/UserPage';
 import { LoginRegister } from './pages/Login/LogInRegister';
-import { isUserLogged } from '../services/authService'
+import { isUserLogged } from '../services/sessionService';
 
 
 class App extends Component {
