+++ src/services/authService.js
@@ -1,7 +1,29 @@
+import { fetchService } from "./fetchService";
+import { REGISTER, LOGIN } from "../shared/constants";
 
-export const isUserLogged= () =>{
-    if(localStorage.getItem("sessionId")){
-        return true
-    }
-    return false
+
+
+ 
+export const postRegister = content => {
+    return fetchService.post(REGISTER, content)
 }
+
+export const postLogin = content => {
+    return fetchService.post(LOGIN, content)
+        .then((response) => {
+            if (response.error) {
+                return response;
+            }
+
+            localStorage.setItem("sessionId", response.sessionId)
+            return response;
+        })
+}
+
+
+ export const logOut = () => {
+    localStorage.removeItem('sessionId');
+}
+
+
+
