+++ src/services/postFetch.js
@@ -1,13 +1,15 @@
 import { fetchService } from "./fetchService";
-import { TEXTPOSTS, VIDEOPOSTS, IMAGEPOST, TEXTPOSTSGET, VIDEOPOSTSGET, IMAGEPOSTSGET, REGISTER, LOGIN } from "../shared/constants";
+import { TEXTPOSTS, VIDEOPOSTS, IMAGEPOST, TEXTPOSTSGET, VIDEOPOSTSGET, IMAGEPOSTSGET } from "../shared/constants";
 
 import { TextPost, VideoPost, ImagePost } from '../entities/Post';
-import { POSTS } from '../shared/constants';
+import { POSTS } from '../shared/constants'
+
 
 
 export const getAllPosts = () => {
     return fetchService.get(POSTS)
         .then((response) => {
+            console.log(response)
             return response.filter(post => {
                 if (post.videoUrl) {
                     return post.videoUrl.includes('youtube')
@@ -15,24 +17,6 @@ export const getAllPosts = () => {
                 return true;
             })
         })
-        .then((response) => {
-            return response.map((post) => {
-                const { videoUrl, imageUrl, text, id, dateCreated, userId, userDisplayName, type, commentsNum } = post;
-                if (type === "text") {
-                    return new TextPost(text, id, dateCreated, userId, userDisplayName, type, commentsNum);
-                }
-                if (type === "image") {
-                    return new ImagePost(imageUrl, id, dateCreated, userId, userDisplayName, type, commentsNum);
-                }
-                if (type === "video") {
-                    return new VideoPost(videoUrl, id, dateCreated, userId, userDisplayName, type, commentsNum);
-                }
-            })
-        })
-}
-
-export const getLastTenPosts = (top, skip) => {
-    return fetchService.get(`${POSTS}?$orderby=DateCreated desc&$top=${top}&$skip=${skip}`)
         .then((response) => {
             return response.map((post) => {
                 const { videoUrl, imageUrl, text, id, dateCreated, userId, userDisplayName, type, commentsNum } = post;
@@ -85,22 +69,6 @@ export const postVideo = content => {
 
 export const postImage = content => {
     return fetchService.post(IMAGEPOST, content)
-}
-
-export const postRegister = content => {
-    return fetchService.post(REGISTER, content)
-}
-
-export const postLogin = content => {
-    return fetchService.post(LOGIN, content)
-        .then((response) => {
-            if (response.error) {
-                return response;
-            }
-
-            localStorage.setItem("sessionId", response.sessionId)
-            return response;
-        })
 }
