+++ src/services/sessionService.js
@@ -1,0 +1,6 @@
+export const isUserLogged= () =>{
+    if(localStorage.getItem("sessionId")){
+        return true
+    }
+    return false
+}
